make ReverseDiff Optional
