+++ Project.toml
@@ -9,7 +9,6 @@ FiniteDiff = "6a86dc24-6348-571c-b903-95158fe2bd41"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
-ReverseDiff = "37e2e3b7-166d-5795-8a7a-e32c996b4267"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 SparseDiffTools = "47a9eef4-7e08-11e9-0b38-333d64bd3804"
 
@@ -19,7 +18,6 @@ FiniteDiff = "2.8"
 ForwardDiff = "0.10"
 Ipopt = "1.0"
 Requires = "1.1"
-ReverseDiff = "1.9"
 SparseDiffTools = "1.13"
 julia = "1.6"
