+++ src/SNOW.jl
@@ -1,10 +1,9 @@
 module SNOW
 
-using ForwardDiff
-using ReverseDiff
+import ForwardDiff
 # using Zygote
 using DiffResults
-using FiniteDiff
+import FiniteDiff
 using SparseArrays
 using SparseDiffTools
 using Requires
@@ -28,6 +27,7 @@ export IPOPT
 # conditionally load Snopt
 function __init__()
     @require Snopt="0e9dc826-d618-11e8-1f57-c34e87fde2c0" include("snopt.jl")
+    @require ReverseDiff = "37e2e3b7-166d-5795-8a7a-e32c996b4267" include("ReverseAD.jl")
 end
 
 end
